requirements.txt
pdfplumber
PyPDF2
reportlab
python-dateutil
PyYAML